requirements.txt
pandas
requests
django
celery
requests
redis
pymongo[srv]
djangorestframework
psycopg2
python-decouple